JS for form.html page
