+++ source/catalog.html
@@ -1258,7 +1258,6 @@
     </footer>
 
     <script src="js/svg4everybody.min.js" onload="svg4everybody({})" async=""></script>
-    <script src="js/main.js"></script>
-    <script src="js/form.js"></script>
+    <script src="js/script.min.js"></script>
   </body>
 </html>
