+++ source/index.html
@@ -516,7 +516,6 @@
     </footer>
 
     <script src="js/svg4everybody.min.js" onload="svg4everybody({})" async=""></script>
-    <script src="js/main.js"></script>
-    <script src="js/index.js"></script>
+    <script src="js/script.min.js"></script>
   </body>
 </html>
