+++ source/sass/blocks/routes.scss
@@ -53,6 +53,45 @@
   &:last-child::before {
     display: none;
   }
+
+  &--active .routes__remove {
+    display: none;
+  }
+
+  &--active .routes__select {
+    color: #ffffff;
+
+    border: none;
+    border-radius: 4px 4px 0 0;
+    background-color: $light-blue;
+
+    &::after,
+    &::before {
+      top: 50%;
+      transform: translateY(-50%) rotate(45deg);
+      right: 9.66%;
+      width: 1px;
+      height: 13px;
+
+      background: none;
+      border: none;
+      border-radius: 0;
+      background-color: #ffffff;
+      opacity: 1;
+    }
+
+    &::after {
+      transform: translateY(-50%) rotate(-45deg);
+    }
+  }
+
+  &--active .routes__text::before {
+    display: none;
+  }
+
+  &--active .routes__alphabeth-wrapper {
+    display: block;
+  }
 }
 
 @media (min-width: $tablet-width) {
@@ -76,6 +115,26 @@
 
       background: linear-gradient(to bottom, #d6d9e6 0%, rgba(214, 217, 230, 0.4) 60%, transparent 100%);
     }
+
+    &--active .routes__remove {
+      display: block;
+    }
+
+    &--active .routes__select {
+      &::after,
+      &::before {
+        width: 2px;
+        height: 25px;
+        right: 36.76%;
+
+        opacity: 1;
+      }
+
+      &:hover::before,
+      &:focus::before {
+        opacity: 0.5;
+      }
+    }
   }
 }
 
@@ -83,6 +142,13 @@
   .routes__item {
     &:last-child::before {
       right: 222px;
+    }
+
+    &--active .routes__select {
+      &::after,
+      &::before {
+        right: 32px;
+      }
     }
   }
 }
@@ -339,4 +405,22 @@
 
 .routes__alphabeth-wrapper {
   display: none;
+  width: 100%;
+  position: absolute;
+  left: 0;
+  top: 42px;
+  z-index: 3;
+}
+
+@media (min-width: $tablet-width) {
+  .routes__alphabeth-wrapper {
+    width: 108.69%;
+    top: 70px;
+  }
+}
+
+@media (min-width: $desktop-width) {
+  .routes__alphabeth-wrapper {
+    width: 730px;
+  }
 }
